cleanup/refactor directory walker
