+++ src/Kit/Io/File/_0test/_0build/linux/gcc-host/output3.txt
@@ -1,0 +1,1 @@
+abob's your unclebobHello Worl !"#$%&'() !"#$%&'()
