+++ src/Kit/Io/File/_0test/_0build/linux/gcc-host/output4.txt
@@ -1,0 +1,3 @@
+
+Hello World
+HelloWorld again!
